change latitude and longitude to double

+++ app/src/main/java/com/cs310/usclassifieds/model/datamodel/Location.java
@@ -2,6 +2,6 @@ package com.cs310.usclassifieds.model.datamodel;
 
 public class Location {
     public String address;
-    public int latitude;
-    public int longitude;
+    public double latitude;
+    public double longitude;
 }
